Added helpers.js
