+++ js_task1/helpers.js
@@ -1,0 +1,3 @@
+export function randomValue(min, max) {
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
