+++ js_task1/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "js_task1",
+  "version": "1.0.0",
+  "description": "",
+  "main": "ATM.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/aotarasenko/TarasenkoAndrey-SummerWeekends.git"
+  },
+  "author": "Tarasenko",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/aotarasenko/TarasenkoAndrey-SummerWeekends/issues"
+  },
+  "homepage": "https://github.com/aotarasenko/TarasenkoAndrey-SummerWeekends#readme",
+  "dependencies": {
+    "babel": "^6.23.0"
+  }
+}
