+++ js_task1/thirdTask.html
@@ -21,6 +21,6 @@
         give out penny
       </div>
     </section>
-    <script src="./thirdTask.js"></script>
+    <script src="./thirdTask.js" type="module"></script>
   </body>
 </html>
